Restructure Node, Multipartycall and Member resources, Update unit tests and example Api's
